requirements.txt
Flask
Jinja2
Flask-SQLAlchemy
factory_boy
